requirements.txt
flask==2.3.3
mysql-connector-python==8.1.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
numpy==1.24.3
pypdf==3.17.0
python-docx==0.8.11